refactor rats.aml module and move a few details to rats.cli
