+++ rats-apps/src/rats/cli/__init__.py
@@ -8,16 +8,21 @@ from ._annotations import (
     get_class_groups,
     group,
 )
-from ._app import ClickApp
+from ._click_app import ClickApp
+from ._command import Command
 from ._container import CompositeContainer, Container
-from ._plugin import attach, create_group
+from ._functions import attach, create_group
+from ._plugin import PluginConfigs, PluginContainer
 
 __all__ = [
     "AUTO_COMMAND",
     "ClickApp",
+    "Command",
     "CommandId",
     "CompositeContainer",
     "Container",
+    "PluginConfigs",
+    "PluginContainer",
     "attach",
     "command",
     "create_group",
