+++ rats-apps/src/rats/cli/_click_app.py
