+++ rats-apps/src/rats/cli/_plugin.py
@@ -1,61 +1,42 @@
-import click
+import os
+import sys
+from collections.abc import Mapping
+from pathlib import Path
+from typing import final
 
-from ._container import Container
+from rats import apps
 
-
-def create_group(group: click.Group, container: Container) -> click.Group:
-    """
-    A simple shortcut to create a [click.Group][] with commands provided by `rats.cli.Container`'s.
-
-    ```python
-    import click
-    from rats import cli
+from ._command import Command
 
 
-    class Example(cli.Container):
-        @cli.command()
-        def my_command(self) -> None:
-            print("hello, world. run me with `my-cli my-command`.")
+@apps.autoscope
+class PluginConfigs:
+    """Service IDs provided and used by [rats.cli.PluginContainer][]."""
+
+    COMMAND = apps.ServiceId[Command]("command")
+    CWD = apps.ServiceId[str]("cwd")
+    ARGV = apps.ServiceId[tuple[str, ...]]("argv")
+    ENV = apps.ServiceId[Mapping[str, str]]("env")
 
 
-    if __name__ == "__main__":
-        cli.create_group(click.Group("my-cli"), Example()).main()
-    ```
+@final
+class PluginContainer(apps.Container, apps.PluginMixin):
+    @apps.service(PluginConfigs.COMMAND)
+    def _command(self) -> Command:
+        return Command(
+            cwd=self._app.get(PluginConfigs.CWD),
+            argv=self._app.get(PluginConfigs.ARGV),
+            env=self._app.get(PluginConfigs.ENV),
+        )
 
-    Args:
-        group: The [click.Group][] commands will be added to.
-        container: The [rats.cli.Container][] instance to attach commands from
-    """
-    container.attach(group)
-    return group
+    @apps.fallback_service(PluginConfigs.CWD)
+    def _cwd(self) -> str:
+        return str(Path.cwd().resolve())
 
+    @apps.fallback_service(PluginConfigs.ARGV)
+    def _argv(self) -> tuple[str, ...]:
+        return tuple(sys.argv)
 
-def attach(
-    group: click.Group,
-    command: click.Command | click.Group,
-    *commands: click.Command | click.Group,
-) -> None:
-    """
-    Convenience function to attach multiple [click.Command]'s to a [click.Group].
-
-    ```python
-    import click
-    from rats import cli
-
-    cli.attach(
-        click.Group("my-cli"),
-        click.Command("command-1"),
-        click.Command("command-2"),
-    )
-    ```
-
-    Args:
-        group: The [click.Group] we want to build.
-        command: A [click.Command] to attach to the provided group.
-        *commands: Any number of additional [click.Command]'s to attach.
-
-    Returns: The input [click.Group][] with all of the commands attached.
-    """
-    group.add_command(command)
-    for c in commands:
-        group.add_command(c)
+    @apps.fallback_service(PluginConfigs.ENV)
+    def _env(self) -> Mapping[str, str]:
+        return dict(os.environ)
