+++ rats-devtools/src/rats/aml/__init__.py
@@ -1,7 +1,8 @@
 """Submit applications as aml jobs, sending necessary context from the driver node."""
 
-from ._app import AppConfigs, Application, AppServices, main, submit
+from ._app import AppConfigs, Application, AppServices, main
 from ._configs import AmlEnvironment, AmlIO, AmlJobContext, AmlJobDetails, AmlWorkspace
+from ._submission import submit
 
 __all__ = [
     "AmlEnvironment",
